apps: move function to manage operations into new file
Move functions that handle mid-level operations management into
a separate source file, as operations.go is simply getting too huge.
